% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibicubic.r
\name{ibicubic}
\alias{ibicubic}
\title{Bicubic interpolation on an array}
\usage{
ibicubic(x, xlim = NULL, ylim = xlim, dx = NULL, dy = dx, par = T)
}
\arguments{
\item{x}{A matrix or higher dimensonal array  containing the
z[i,j] data values for the grid points (x[i],y[j]).}

\item{xlim}{Lower and upper limits (x1, x2) for x-coordinate used for output.}

\item{ylim}{Lower and upper limits (y1, y2) for y-coordinate used for output.}

\item{dx}{Output grid spacing in x direction.}

\item{dy}{Output grid spacing in y direction.}

\item{par}{Use \link{parallel} package to interpolate.}
}
\value{
Interpolated array
}
\description{
First two dimensions of the array is interpolated by bicubic grid.
Other dimensions are repeated. If \code{xlim} and \code{ylim} are
not set, range of first two dimensions are used. If \code{dx} and
\code{dy} are not set, half of the grid spacing is used. If \code{x}
matrix/array has \code{\link{dimnames}}, they are used as coordinates
of the rectangular grid.
}
\examples{
x <- array(rnorm(1000), dim = c(10, 10, 10))
xi <- ibicubic(x)
str(xi) # should have dim = c(19, 19, 10)

x <- array(rnorm(1000), dim = c(10, 10, 10),
           dimnames = list(seq(2, 20, 2), seq(2, 20, 2), NULL))
xi <- ibicubic(x)
str(xi) # note dimnames

xi <- ibicubic(x, xlim = c(4, 6), ylim = c(6, 8))
str(xi) # note dimnames

}

