% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsezen.r
\docType{package}
\name{rsezen}
\alias{rsezen}
\alias{rsezen-package}
\title{rsezen: A package for processing data.}
\description{
The rsezen package provides the functions below:
\code{\link{ibicubic}}.
}
\section{rsezen functions}{

The rwind functions ...
}

