% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsezen.r
\docType{package}
\name{rsezen}
\alias{rsezen}
\alias{rsezen-package}
\title{rwind: A package for processing wind and related data.}
\description{
The rwind package provides four categories of important functions:
\code{\link{deg2comp}}, \code{\link{hd2uv}}, \code{\link{uv2hd}}
and \code{\link{rotax}}.
}
\section{rsezen functions}{

The rwind functions ...
}

